// frontend/src/enums/buy.enum.ts
export enum BuyOccupation {
  SALARIED = 'salaried',
  SELF_EMPLOYED = 'self-employed',
  RETIRED = 'retired',
  STUDENT = 'student',
}

export enum BuyEducation {
  POST_GRADUATE_AND_ABOVE = 'postgraduate',
  TWELFTH = '12th',
  TENTH = '10th',
  BELOW_TENTH = 'below10th',
  GRADUATE = 'graduate',
}

export enum BuyNationality {
  RESIDENT_INDIAN = 'residentIndian',
}

export enum BuyMedicalHistory {
  YES = 'yes',
  NO = 'no',
}

export enum BuyPlanOption {
  ICRORE = 'icrore',
  TEN_LAKH = '10lakh',
  TEN_LAKH_PLUS = '10lakhplus',
}

export enum BuyPayFor {
  ONE_TIME = 'onetime',
  FIVE_YEARS = '5years',
  TEN_YEARS = '10years',
  TWENTY_YEARS = '20years',
}

export enum BuyPaymentMode {
  MONTHLY = 'monthly',
  YEARLY = 'yearly',
}

export enum BuyPaymentMethod {
  DEBIT = 'debit',
  CREDIT = 'credit',
  NETBANKING = 'netbanking',
  UPI = 'upi',
  WALLET = 'wallet',
}